+++ src/mods/AdventureGuide/src/Resolution/NavigationTargetResolver.cs
@@ -24,7 +24,7 @@ public sealed class NavigationTargetResolver
 	private int _lastBatchKeyCount;
 	private IReadOnlyList<QuestCostSample> _topQuestCosts = Array.Empty<QuestCostSample>();
 
-	public int Version => _reader.Engine.Revision;
+	public long Version => _reader.Engine.Revision;
 
 	internal NavigationTargetResolver(
 		CompiledGuideModel guide,
